third_party/blink/web_tests/fast/dom/MutationObserver/observe-options-character-data-expected.txt

Test MutationObserver.observe options related to character data

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS records.length is 1
PASS records[0].oldValue is "0"
PASS records.length is 1
PASS records[0].oldValue is null
PASS successfullyParsed is true

TEST COMPLETE
